refactored for activity robot
